requirements.txt
flask
flask_httpauth
flask-jwt-extended
pymongo
python-dotenv
ping3